partial refactor of GpuFileSourceScanExec
